+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.x branch
 ## 2.22 branch
 ### 2.22.3
-* fixed a regression for `FormHelper::createInline()` introduced with the `2.22.2` version.
+* fixed a regression for `FormHelper::createInline()` introduced with the `2.22.2` version;
+* fixed the class of submit buttons in inline forms.
 
 ### 2.22.2
 * `assertExitWithError()` and `assertExitWithSuccess()` methods provided by `ConsoleIntegrationTestTrait` are now